requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
playwright==1.44.0
httpx==0.27.0
python-multipart==0.0.9
google-cloud-storage==2.16.0
google-auth==2.30.0
pydantic==2.11.7